rename and remove comments
